Fix default network gateway detection
