+++ internal/vpn/os.go
@@ -1,7 +1,6 @@
 package vpn
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"net"
@@ -13,43 +12,6 @@ import (
 // SetupTUN sets the allocated TUN interface up, setting its IP, gateway, netmask and MTU.
 func SetupTUN(ifcName, ip, netmask, gateway string, mtu int) error {
 	return run("ifconfig", ifcName, ip, gateway, "mtu", strconv.Itoa(mtu), "netmask", netmask, "up")
-}
-
-// NetworkInterfaceGateway gets gateway of the network interface with name `ifcName`.
-func NetworkInterfaceGateway(ifcName string) (net.IP, error) {
-	cmd := fmt.Sprintf(gatewayForIfcCMDFmt, ifcName)
-	outBytes, err := exec.Command("sh", "-c", cmd).Output() //nolint:gosec
-	if err != nil {
-		return nil, fmt.Errorf("error running command %s: %w", cmd, err)
-	}
-
-	outBytes = bytes.TrimRight(outBytes, "\n")
-
-	outLines := bytes.Split(outBytes, []byte{'\n'})
-
-	for _, l := range outLines {
-		if bytes.Count(l, []byte{'.'}) != 3 {
-			// initially look for IPv4 address
-			continue
-		}
-
-		ip := net.ParseIP(string(l))
-		if ip != nil {
-			return ip, nil
-		}
-	}
-
-	return nil, fmt.Errorf("couldn't find gateway IP for \"%s\"", ifcName)
-}
-
-// DefaultNetworkGateway fetches system's default network gateway.
-func DefaultNetworkGateway() (net.IP, error) {
-	defaultNetworkIfcName, err := DefaultNetworkInterface()
-	if err != nil {
-		return nil, fmt.Errorf("error getting default network interface name: %w", err)
-	}
-
-	return NetworkInterfaceGateway(defaultNetworkIfcName)
 }
 
 // DefaultNetworkInterface fetches default network interface name.
