+++ internal/vpn/os_darwin.go
@@ -6,6 +6,8 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"net"
+	"os/exec"
 )
 
 const (
@@ -15,6 +17,16 @@ const (
 	getIPv4ForwardingCMD    = "sysctl net.inet.ip.forwarding"
 	getIPv6ForwardingCMD    = "sysctl net.inet6.ip6.forwarding"
 )
+
+// DefaultNetworkGateway fetches system's default network gateway.
+func DefaultNetworkGateway() (net.IP, error) {
+	defaultNetworkIfcName, err := DefaultNetworkInterface()
+	if err != nil {
+		return nil, fmt.Errorf("error getting default network interface name: %w", err)
+	}
+
+	return networkInterfaceGateway(defaultNetworkIfcName)
+}
 
 // EnableIPMasquerading enables IP masquerading for the interface with name `ifcName`.
 func EnableIPMasquerading(_ string) error {
@@ -42,6 +54,33 @@ func DeleteRoute(ip, gateway, netmask string) error {
 	}
 
 	return run("route", "delete", "-net", ip, gateway, netmask)
+}
+
+// networkInterfaceGateway gets gateway of the network interface with name `ifcName`.
+func networkInterfaceGateway(ifcName string) (net.IP, error) {
+	cmd := fmt.Sprintf(gatewayForIfcCMDFmt, ifcName)
+	outBytes, err := exec.Command("sh", "-c", cmd).Output() //nolint:gosec
+	if err != nil {
+		return nil, fmt.Errorf("error running command %s: %w", cmd, err)
+	}
+
+	outBytes = bytes.TrimRight(outBytes, "\n")
+
+	outLines := bytes.Split(outBytes, []byte{'\n'})
+
+	for _, l := range outLines {
+		if bytes.Count(l, []byte{'.'}) != 3 {
+			// initially look for IPv4 address
+			continue
+		}
+
+		ip := net.ParseIP(string(l))
+		if ip != nil {
+			return ip, nil
+		}
+	}
+
+	return nil, fmt.Errorf("couldn't find gateway IP for \"%s\"", ifcName)
 }
 
 func parseIPForwardingOutput(output []byte) (string, error) {
